overlay: add -n option (no firstframe)

+++ overlay/overlay.zsh
@@ -2,10 +2,11 @@
 source ~/.zshrc.avenc
 iargs=()
 oargs=()
-while getopts "s:t:" o; do
+while getopts "s:t:n" o; do
     case $o in
         s) iargs=(-ss $OPTARG) ;;
         t) oargs=(-t $OPTARG); length=$OPTARG;;
+	n) nofirstframe=1
     esac
 done
 shift $((OPTIND-1))
@@ -19,5 +20,8 @@ for pass in 1 2
             [v0] pad=600:446:20:35:black [v1],
             [v1] [1:v] overlay [v2]' \
         -map '[v2]' -map '0:a' -ac 2 -crf 20 -b:v $(hkbr $length)  -pass $pass -y "$(basename $1)-overlay-v.webm" || exit 1
-mkvmerge thread-firstframe.webm + "$(basename $1)-overlay-v.webm" -o "$(basename $1)-overlay.webm" || exit 1
-rm "$(basename $1)-overlay-v.webm"
+if ! ((nofirstframe))
+then
+    mkvmerge thread-firstframe.webm + "$(basename $1)-overlay-v.webm" -o "$(basename $1)-overlay.webm" || exit 1
+    rm "$(basename $1)-overlay-v.webm"
+fi
